Correct typr of index

+++ rust-tcp-latency/src/connection/mod.rs
@@ -35,7 +35,7 @@ pub fn receive_message(n_bytes: usize, stream: &mut TcpStream, rbuf: &mut Vec<u8
     }
 }
 
-pub fn pin_thread(index: u8) {
+pub fn pin_thread(index: usize) {
     let core_ids = core_affinity::get_core_ids().unwrap();
     core_affinity::set_for_current(core_ids[index % core_ids.len()]);
 }
